refactor: move schema to its own folders
