+++ src/lib/components/hall/components/CreateRoom.tsx
@@ -4,12 +4,9 @@ import { useRouter } from 'next/router';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 
-import {
-  createRoomFormSchema,
-  initialValues,
-} from '~/lib/components/hall/schema';
 import SpokerInput from '~/lib/components/shared/SpokerInput';
 import SpokerWrapperGrid from '~/lib/components/shared/SpokerWrapperGrid';
+import { createRoomFormSchema, initialValues } from '~/lib/models/hall';
 import { createRoom } from '~/lib/services/firebase/room/create';
 import { formatId } from '~/lib/utils/formatId';
 import { removeFirebasePrefix } from '~/lib/utils/removeFirebasePrefix';
